Aggregate shared API types
