+++ packages/api/src/services/category-service.ts
@@ -1,20 +1,11 @@
-import type { Category, PaginatedResult } from "@listee/types";
 import type {
-  CategoryRepository,
+  Category,
+  CategoryService,
+  CategoryServiceDependencies,
   FindCategoryRepositoryParams,
   ListCategoriesRepositoryParams,
-} from "../repositories/category-repository";
-
-export interface CategoryService {
-  listByUserId(
-    params: ListCategoriesRepositoryParams,
-  ): Promise<PaginatedResult<Category>>;
-  findById(params: FindCategoryRepositoryParams): Promise<Category | null>;
-}
-
-export interface CategoryServiceDependencies {
-  readonly repository: CategoryRepository;
-}
+  PaginatedResult,
+} from "@listee/types";
 
 export function createCategoryService(
   dependencies: CategoryServiceDependencies,
