+++ packages/auth/package.json
@@ -14,6 +14,6 @@
     "clean": "rm -rf dist"
   },
   "dependencies": {
-    "@listee/db": "workspace:^"
+    "@listee/types": "workspace:^"
   }
 }
